Move misc utils to `core`
