+++ divik/__init__.py
@@ -1,17 +1,14 @@
 __version__ = '2.3.17'
 
-from ._seeding import seeded
-from ._utils import DivikResult
+from divik import core
 from divik import feature_selection
 from divik import feature_extraction
 from divik import cluster
 from divik import sampler
 from ._summary import plot, reject_split
-from ._gin_compat import (
-    configurable,
-    parse_gin_args
-)
 
+
+from .core import configurable
 for __estimator in [
     feature_extraction.KneePCA,
     feature_extraction.LocallyAdjustedRbfSpectralEmbedding,
@@ -32,14 +29,13 @@ for __estimator in [
 ]:
     configurable(__estimator)
 
+
 __all__ = [
     "__version__",
+    "core",
     "cluster",
     "feature_selection",
     "feature_extraction",
     "sampler",
-    "seeded",
-    "configurable", "parse_gin_args",
-    'DivikResult',
     "plot", "reject_split",
 ]
