+++ divik/_cli/divik.py
@@ -12,7 +12,7 @@ from divik._cli._data_io import DIVIK_RESULT_FNAME
 from divik.cluster import DiviK
 import divik._summary as _smr
 import divik._cli._utils as sc
-import divik._utils as u
+import divik.core as u
 
 
 def _make_summary(result: typing.Optional[u.DivikResult]):
