+++ divik/cluster/_kmeans/_core.py
@@ -9,7 +9,7 @@ from divik.cluster._kmeans._initialization import \
     Initialization, \
     ExtremeInitialization, \
     PercentileInitialization
-from divik._utils import (
+from divik.core import (
     normalize_rows,
     Centroids,
     IntLabels,
