+++ divik/feature_selection/_gmm_selector.py
@@ -71,9 +71,9 @@ class GMMSelector(BaseEstimator, StatSelectorMixin):
 
     Examples
     --------
-    >>> import numpy as np
+import divik._utils    >>> import numpy as np
     >>> import divik.feature_selection as fs
-    >>> np.random.seed(42)
+    >>> divik._utils.seed(42)
     >>> labels = np.concatenate([30 * [0] + 20 * [1] + 30 * [2] + 40 * [3]])
     >>> data = labels * 5 + np.random.randn(*labels.shape)
     >>> fs.GMMSelector('mean').fit_transform(data)
