+++ docs/modules/divik.core.rst
@@ -1,0 +1,7 @@
+`core` module
+================
+
+.. automodule:: divik.core
+   :members:
+   :undoc-members:
+   :inherited-members:
